Autenticacion1718G2/requirements.txt
django==1.11
djangorestframework==3.7.3
django-rest-auth==0.9.2
django-allauth==0.34.0
mysqlclient==1.3.12
numpy==1.13.3
pytz==2017.3